server failure detection on the datapath is too complex
for example, suppose a client sends a single-packet request to a server
that's running an hour-long request handler in the background. in such
a case, the server will simply drop the client's retransmissions, and
the client will believe the server to be failed. fixing this will
require adding a new datapath packet type, which is a no-go.

it's better to detect server failures on the management path, but this
time without ENet.